Start work on Atlan POC
This initial commit creates some new data structures and
adds to existing OpenLineageClient allowing transmission
to the Atlan endpoint.

The point of this POC is not to make a lasting connection
to Atlan by using their API, but to start to figure out
how to effectively send to multiple endpoints and
to understand what data is missing that Atlan would need
to properly ingest events sent from OpenLineage. Ideally,
our partnership with Atlan will lead to some integration
on their end, so they can ingest a RunEvent and convert
it to their "Process" internally.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>